requirements.txt
setuptools<81
bert_score==0.3.13
bpy==4.4.0
einops==0.8.1
hydra-core==1.3.2
imageio==2.36.1
ipython==8.31.0
joblib==1.3.2
gdown==5.2.0
matplotlib==3.10.3
moviepy==1.0.3
nlg_metricverse==0.9.9
numpy==1.26.0
omegaconf==2.3.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
orjson==3.10.18
pandas==2.3.0
peft==0.15.2
Pillow==11.2.1
psutil==6.1.1
pyrender==0.1.45
pytorch_lightning==2.0.0
rich==14.0.0
scikit_learn==1.7.0
scipy==1.15.3
Shapely==2.1.1
shortuuid==1.0.13
spacy==3.8.3
thinc==8.3.3
torchmetrics
tqdm==4.67.1
transformers==4.47.1
trimesh==3.9.24
wandb
